updated inheritance
