+++ mvlearn/cluster/base.py
@@ -31,8 +31,47 @@ class BaseCluster(BaseEstimator):
         pass
 
     @abstractmethod
-    def fit_predict(self, Xs, y=None):
+    def fit(self, Xs, y=None):
+        '''
+        A method to fit clustering parameters to the multiview data.
+        Parameters
+        ----------
+        Xs : list of array-likes or numpy.ndarray
+            - Xs length: n_views
+            - Xs[i] shape: (n_samples, n_features_i)
+            A list of different views to fit the model on.
 
+        y : array-like, shape (n_samples,)
+            Labels for each sample. Only used by supervised algorithms.
+
+        Returns
+        -------
+        self :  returns and instance of self.
+        '''
+
+        return self
+
+    @abstractmethod
+    def predict(self, Xs):
+        '''
+        A method to predict cluster labels of multiview data.
+        Parameters
+        ----------
+        Xs : list of array-likes or numpy.ndarray
+            - Xs length: n_views
+            - Xs[i] shape: (n_samples, n_features_i)
+            A list of different views to cluster.
+
+        Returns
+        -------
+        labels : array-like, shape (n_samples,)
+            Returns the predicted cluster labels for each sample.
+        '''
+
+        labels = self.labels_
+        return labels
+
+    def fit_predict(self, Xs, y=None):
         '''
         A method for fitting then predicting cluster assignments.
 
@@ -41,6 +80,7 @@ class BaseCluster(BaseEstimator):
         Xs : list of array-likes or numpy.ndarray
             - Xs length: n_views
             - Xs[i] shape: (n_samples, n_features_i)
+            A list of different views to fit the model on.
 
         y : array-like, shape (n_samples,)
             Labels for each sample. Only used by supervised algorithms.
@@ -51,4 +91,6 @@ class BaseCluster(BaseEstimator):
             The predicted cluster labels for each sample.
         '''
 
-        pass
+        self.fit(Xs)
+        labels = self.labels_
+        return labels
