+++ tests/cluster/test_coreg.py
