+++ tests/cluster/test_kmeans.py
