+++ tests/cluster/test_spectral.py
