+++ tests/cluster/test_spherical_kmeans.py
